added validators

+++ src/comments/comments.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/sequelize';
 import { Post } from 'src/posts/posts.model';
 import { User } from 'src/users/users.model';
@@ -41,9 +41,13 @@ export class CommentsService {
     });
   }
 
-  async newComment(dto: CreateCommenttDto) {
+  async newComment(dto: CreateCommenttDto, req) {
     try {
-      const newComment = await this.commentsRepository.create(dto);
+      const userId = req.user.id;
+      const newComment = await this.commentsRepository.create({
+        ...dto,
+        userId,
+      });
       const user = await newComment.$get('user', {
         attributes: ['id', 'fullName', 'email', 'avatarUrl'],
       });
@@ -56,13 +60,22 @@ export class CommentsService {
     }
   }
 
-  async deleteComment(id: number) {
-    const result = await this.commentsRepository.destroy({
-      where: { id },
-    });
-    if (result) {
-      return { success: true };
+  async deleteComment(id: number, req) {
+    try {
+      const comment = await this.commentsRepository.findByPk(id);
+      if (!comment) return;
+      if (comment.dataValues.userId !== req.user.id) {
+        return new UnauthorizedException();
+      }
+      const result = await this.commentsRepository.destroy({
+        where: { id },
+      });
+      if (result) {
+        return { success: true };
+      }
+      return;
+    } catch (error) {
+      throw error;
     }
-    return;
   }
 }
